Assignment07 08

+++ infrastructure/aws/cloudformation/csye6225-aws-cf-terminate-application-stack.sh
@@ -1,24 +1,30 @@
 STACK_NAME=$1
 EC2_NAME=${STACK_NAME}-csye6225-ec2
 
-bucketName=$(aws s3api list-buckets --query "Buckets[].Name" --output text | grep me.csye6225.com)
+bucketName=$(aws s3api list-buckets --query "Buckets[].Name" --output text | grep me.csye6225.com | awk '{print $1}')
 echo $bucketName
-aws s3 rm s3://$bucketName --recursive
+if [ -n "$bucketName" ]
+then
+    aws s3 rm s3://$bucketName --recursive
+fi
 
 export ec2InstanceId=$(aws ec2 describe-instances --query 'Reservations[*].Instances[*].[InstanceId, State.Name, Tags[*][?Value==`${EC2_NAME}`]]' --output text|grep running|awk '{print $1}')
+if [ -n "$ec2InstanceId" ]
+then
+	aws ec2 stop-instances --instance-ids $ec2InstanceId
+	aws ec2 wait instance-stopped --instance-ids $ec2InstanceId
+	echo "Instance ${ec2InstanceId} is stopped!!"
 
-aws ec2 stop-instances --instance-ids $ec2InstanceId
-aws ec2 wait instance-stopped --instance-ids $ec2InstanceId
-echo "Instance ${ec2InstanceId} is stopped!!"
-
-aws ec2 modify-instance-attribute --no-disable-api-termination --instance-id $ec2InstanceId
-echo "Termination protection is removed from the instance!!"
+	aws ec2 modify-instance-attribute --no-disable-api-termination --instance-id $ec2InstanceId
+	echo "Termination protection is removed from the instance!!"
+fi
 
 aws cloudformation delete-stack --stack-name $STACK_NAME
 aws cloudformation wait stack-delete-complete --stack-name $STACK_NAME
 
-if [ $? -eq 0 ]; then
-echo "Stack ${STACK_NAME} deleted successfully!!"
+if [ $? -eq 0 ]
+then
+	echo "Stack ${STACK_NAME} deleted successfully!!"
 else
-echo "Unable to delete stack. Please input correct name!!"
+	echo "Unable to delete stack. Please input correct name!!"
 fi
